Add ReplayFromDumpTestBase to avoid tests to execute multi times
