+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/ReplayFromDumpTest.java
@@ -14,14 +14,8 @@
 
 package com.starrocks.sql.plan;
 
-import com.google.common.collect.Lists;
-import com.starrocks.common.Config;
 import com.starrocks.common.FeConstants;
 import com.starrocks.common.Pair;
-import com.starrocks.common.io.Writable;
-import com.starrocks.persist.EditLog;
-import com.starrocks.persist.gson.GsonUtils;
-import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.SessionVariable;
 import com.starrocks.qe.VariableMgr;
 import com.starrocks.sql.optimizer.OptExpression;
@@ -30,158 +24,14 @@ import com.starrocks.sql.optimizer.base.CTEProperty;
 import com.starrocks.sql.optimizer.dump.QueryDumpInfo;
 import com.starrocks.sql.optimizer.rule.RuleSet;
 import com.starrocks.sql.optimizer.rule.transformation.JoinAssociativityRule;
-import com.starrocks.system.BackendCoreStat;
 import com.starrocks.thrift.TExplainLevel;
-import com.starrocks.utframe.StarRocksAssert;
 import com.starrocks.utframe.UtFrameUtils;
 import mockit.Mock;
 import mockit.MockUp;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.List;
-import java.util.Objects;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-public class ReplayFromDumpTest {
-    public static ConnectContext connectContext;
-    public static StarRocksAssert starRocksAssert;
-
-    public static List<String> MODEL_LISTS = Lists.newArrayList("[end]", "[dump]", "[result]", "[fragment]",
-            "[fragment statistics]");
-
-    @BeforeClass
-    public static void beforeClass() throws Exception {
-        UtFrameUtils.createMinStarRocksCluster();
-        // Should disable Dynamic Partition in replay dump test
-        Config.dynamic_partition_enable = false;
-        Config.tablet_sched_disable_colocate_overall_balance = true;
-        // create connect context
-        connectContext = UtFrameUtils.createDefaultCtx();
-        connectContext.getSessionVariable().setOptimizerExecuteTimeout(30000);
-        connectContext.getSessionVariable().setJoinImplementationMode("auto");
-        connectContext.getSessionVariable().setCboPushDownAggregateMode(-1);
-        starRocksAssert = new StarRocksAssert(connectContext);
-        FeConstants.runningUnitTest = true;
-        FeConstants.showLocalShuffleColumnsInExplain = false;
-
-        new MockUp<EditLog>() {
-            @Mock
-            protected void logEdit(short op, Writable writable) {
-                return;
-            }
-        };
-    }
-
-    @Before
-    public void before() {
-        BackendCoreStat.reset();
-        connectContext.getSessionVariable().setCboPushDownAggregateMode(-1);
-    }
-
-    @AfterClass
-    public static void afterClass() throws Exception {
-        connectContext.getSessionVariable().setEnableLocalShuffleAgg(true);
-        FeConstants.showLocalShuffleColumnsInExplain = true;
-    }
-
-    public String getModelContent(String filename, String model) {
-        String path = Objects.requireNonNull(ClassLoader.getSystemClassLoader().getResource("sql")).getPath();
-        File file = new File(path + "/" + filename + ".sql");
-        StringBuilder modelContentBuilder = new StringBuilder();
-
-        String tempStr;
-        boolean hasModelContent = false;
-        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
-            while ((tempStr = reader.readLine()) != null) {
-                if (tempStr.equals("[" + model + "]")) {
-                    hasModelContent = true;
-                    continue;
-                }
-
-                if (MODEL_LISTS.contains(tempStr)) {
-                    hasModelContent = false;
-                }
-
-                if (hasModelContent) {
-                    modelContentBuilder.append(tempStr).append("\n");
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            Assert.fail();
-        }
-        return modelContentBuilder.toString();
-    }
-
-    public QueryDumpInfo getDumpInfoFromJson(String dumpInfoString) {
-        return GsonUtils.GSON.fromJson(dumpInfoString, QueryDumpInfo.class);
-    }
-
-    public SessionVariable getTestSessionVariable() {
-        SessionVariable sessionVariable = VariableMgr.newSessionVariable();
-        sessionVariable.setMaxTransformReorderJoins(8);
-        sessionVariable.setEnableGlobalRuntimeFilter(true);
-        sessionVariable.setEnableMultiColumnsOnGlobbalRuntimeFilter(true);
-        return sessionVariable;
-    }
-
-    public void compareDumpWithOriginTest(String filename) throws Exception {
-        String dumpString = getModelContent(filename, "dump");
-        String originCostPlan = Stream.of(getModelContent(filename, "fragment statistics").split("\n")).
-                filter(s -> !s.contains("tabletList")).collect(Collectors.joining("\n"));
-        String replayCostPlan = Stream.of(
-                        PlanTestBase.format(getCostPlanFragment(dumpString, getTestSessionVariable()).second).split("\n"))
-                .filter(s -> !s.contains("tabletList")).collect(Collectors.joining("\n"));
-        Assert.assertEquals(originCostPlan, replayCostPlan);
-    }
-
-    protected String getDumpInfoFromFile(String fileName) throws Exception {
-        String path = Objects.requireNonNull(ClassLoader.getSystemClassLoader().getResource("sql")).getPath();
-        File file = new File(path + "/" + fileName + ".json");
-        StringBuilder sb = new StringBuilder();
-        String tempStr;
-        try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
-            while ((tempStr = reader.readLine()) != null) {
-                sb.append(tempStr);
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-            Assert.fail();
-        }
-
-        return sb.toString();
-    }
-
-    private Pair<QueryDumpInfo, String> getCostPlanFragment(String dumpJonStr) throws Exception {
-        return getCostPlanFragment(dumpJonStr, null);
-    }
-
-    protected Pair<QueryDumpInfo, String> getCostPlanFragment(String dumpJsonStr, SessionVariable sessionVariable)
-            throws Exception {
-        return getPlanFragment(dumpJsonStr, sessionVariable, TExplainLevel.COSTS);
-    }
-
-    protected Pair<QueryDumpInfo, String> getPlanFragment(String dumpJsonStr, SessionVariable sessionVariable,
-                                                          TExplainLevel level) throws Exception {
-        QueryDumpInfo queryDumpInfo = getDumpInfoFromJson(dumpJsonStr);
-        if (sessionVariable != null) {
-            queryDumpInfo.setSessionVariable(sessionVariable);
-        }
-        queryDumpInfo.getSessionVariable().setOptimizerExecuteTimeout(30000000);
-        queryDumpInfo.getSessionVariable().setCboPushDownAggregateMode(-1);
-        return new Pair<>(queryDumpInfo,
-                UtFrameUtils.getNewPlanAndFragmentFromDump(connectContext, queryDumpInfo).second.
-                        getExplainString(level));
-    }
+public class ReplayFromDumpTest extends ReplayFromDumpTestBase {
 
     @Test
     public void testTPCH1() throws Exception {
