+++ fe/fe-core/src/test/java/com/starrocks/sql/plan/ReplayWithMVFromDumpTest.java
@@ -29,11 +29,11 @@ import org.junit.Test;
 
 import java.util.Set;
 
-public class ReplayWithMVFromDumpTest extends ReplayFromDumpTest {
+public class ReplayWithMVFromDumpTest extends ReplayFromDumpTestBase {
 
     @BeforeClass
     public static void beforeClass() throws Exception {
-        ReplayFromDumpTest.beforeClass();
+        ReplayFromDumpTestBase.beforeClass();
         connectContext.getSessionVariable().setEnableMVOptimizerTraceLog(true);
 
         new MockUp<MaterializedView>() {
